% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{use_tippy}
\alias{tippy_class}
\alias{with_tippy}
\title{Use tippy}
\usage{
use_tippy()

tippy_class(class, ...)

with_tippy(element, tooltip, ...)
}
\arguments{
\item{class}{Class of elements to apply tooltip to.}

\item{...}{Any option from \href{https://atomiks.github.io/tippyjs/#all-options}{the official documentation}.}

\item{element}{Shiny, or htmltools element, or character string.}

\item{tooltip}{Content of tooltip.}
}
\description{
__Deprecated.__ Dynamically use tippy.
}
\section{Functions}{

\itemize{
  \item{\code{use_tippy} Includes tippy.js in header.}
  \item{\code{call_tippy} call \code{tippy} on specific target element(s), place after elements to be targeted.}
}
}

\examples{
library(shiny)

if(FALSE){
shinyApp(
  ui = fluidPage(
    p("Some text", class = "tooltip"),
    p("Some text", class = "tooltip"),
    p("Some text", class = "tooltip"),
    p("Some text", class = "tooltip"),
    p("Some text", class = "tooltip"),
    p("Some text", class = "tooltip"),
    tippy("tooltip", content = "Hi!") # all elements with class
 ),
 server = function(input, output) {}
)
}

}
\seealso{
\href{https://atomiks.github.io/tippyjs/}{official documentation}
}
