% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_selector}
\alias{is_selector}
\title{Is Selector}
\usage{
is_selector(selector)
}
\arguments{
\item{selector}{Selector to check.}
}
\description{
Check whether the input is a valid selector
assuming it's a tag selector.
}
\keyword{internal}
