% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_selector}
\alias{make_selector}
\title{Make Selector}
\usage{
make_selector(selector, .is_tag = FALSE)
}
\arguments{
\item{selector}{Selector.}

\item{.is_tag}{Do not pre-process selector.}
}
\value{
Valid selector.
}
\description{
Make Selector
}
\keyword{internal}
