% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/tippy.R
\name{with_tippy}
\alias{with_tippy}
\alias{tippy}
\title{Add a Tooltip to an Element}
\usage{
with_tippy(element, tooltip, ...)

tippy(element, content, ...)
}
\arguments{
\item{element}{Shiny, or htmltools element, or character string.}

\item{...}{Any other options from \href{https://atomiks.github.io/tippyjs/v6/all-props/}{the official documentation}.}

\item{content}{Content of the tooltip.}
}
\description{
Add tooltips to an element.
}
\examples{
library(shiny)

tippy("Hover me!", content = "Hi, I'm the tooltip!")
tippy(h3("Hello"), content = "World")

}
\seealso{
\href{https://atomiks.github.io/tippyjs/}{official documentation}
}
