% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{tippy_global_theme}
\alias{tippy_global_theme}
\title{Theme}
\usage{
tippy_global_theme(name)
}
\arguments{
\item{name}{Name of the theme to set.}
}
\description{
Set a global theme. 
Every tooltip will subsequently use this theme.
}
