% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tippy_plot}
\alias{tippy_plot}
\title{Use a plot as tooltip}
\usage{
tippy_plot(
  selector,
  plotId,
  ...,
  .width = 200,
  .height = 200,
  .hidePlot = FALSE,
  .session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{selector}{A CSS selector, e.g.: `#id` or
`.class`. If it is a bare selector (e.g.: `sth`)
then it is assumed to be an id and processed as
`#sth`. Set `.is_tag` to `TRUE` to disable that.}

\item{plotId}{Id of plot to use.}

\item{...}{Any other options from \href{https://atomiks.github.io/tippyjs/v6/all-props/}{the official documentation}.}

\item{.width, .height}{Dimensions of plot.}

\item{.hidePlot}{Whether to hide the original plot.}

\item{.session}{A valid shiny session.}
}
\description{
Use a plot as tooltip
}
\examples{
library(shiny)

ui <- fluidPage(
  useTippy(),
  plotOutput("plot"),
  h5("plot tooltip", id = "plotTip")
)

server <- function(input, output) {
  output$plot <- renderPlot({
    on.exit({
      tippy_plot(
        "plotTip", 
        "plot", 
        theme = "light",
        .hidePlot = TRUE,
				 .width = 600,
				 .height = 200,
        maxWidth = 700
      )
    })
    plot(cars)
  })
}

if(FALSE){
  shinyApp(ui, server)
}

}
