% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{tippyDefault}
\alias{tippyDefault}
\title{Default}
\usage{
tippyDefault(...)
}
\arguments{
\item{...}{Any other options from \href{https://atomiks.github.io/tippyjs/v6/all-props/}{the official documentation}.}
}
\description{
Set tippy defaults that will be shared
by all other tooltips.
}
