% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{tippyTheme}
\alias{tippyTheme}
\title{Theme}
\usage{
tippyTheme(
  name,
  box_bg = "black",
  box_color = "white",
  box_font_size = "12px",
  arrow_color = box_bg,
  content_bg = box_bg,
  content_color = box_color,
  content_font_size = box_font_size
)
}
\arguments{
\item{name}{Name of theme.}

\item{box_bg, box_color, content_bg, content_color}{Background and text color respectively.}

\item{box_font_size, content_font_size}{Size of the font.}

\item{arrow_color}{Color of the arrow (if enabled).}
}
\description{
Convenience function to create a theme.
}
\examples{
library(shiny)

ui <- fluidPage(
  useTippy(),
  tippyTheme("myTheme", box_bg = "darkred"),
  h5("plot tooltip", id = "plotTip"),
  tippyThis(
    "plotTip",
    "Hello world",
    theme = "myTheme"
  )
)

server <- function(input, output) {}

if(FALSE){
  shinyApp(ui, server)
}

}
