% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/tippy.R
\name{tippy_this}
\alias{tippy_this}
\alias{tippyThis}
\title{Bind Tooltip to a Selector}
\usage{
tippy_this(selector, content = NULL, ...)

tippyThis(selector, content = NULL, ..., .is_tag = FALSE)
}
\arguments{
\item{selector}{A CSS selector, e.g.: `#id` or
`.class`. If it is a bare selector (e.g.: `sth`)
then it is assumed to be an id and processed as
`#sth`. Set `.is_tag` to `TRUE` to disable that.}

\item{content}{Content of the tooltip.}

\item{...}{Any other options from \href{https://atomiks.github.io/tippyjs/v6/all-props/}{the official documentation}.}

\item{.is_tag}{Whether the selector is a tag,
e.g.: `<p>`.}
}
\description{
Binds a tooltip to a valid CSS selector.
}
\examples{
library(shiny)

ui <- fluidPage(
 tippy(h3("Hello"), "World"),
 h4("World", id = "theId"),
 tippyThis("theId", "A tooltip")
)

server <- function(input, output) {}

if(interactive())
 shinyApp(ui, server)

}
