% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{tippyTheme}
\alias{tippyTheme}
\title{Theme}
\usage{
tippyTheme(...)
}
\arguments{
\item{...}{Any other options from \href{https://atomiks.github.io/tippyjs/v6/all-props/}{the official documentation}.}
}
\description{
Set tippy theme.
}
